pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "academic-document-manager"
version = "1.0.0"
description = "Desktop application for academic document processing"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ADM Team", email = "adm@example.com"}
]
keywords = ["academic", "document", "thesis", "latex", "docx", "pdf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
requires-python = ">=3.10"
dependencies = [
    "python-docx>=0.8.11",
    "PyMuPDF>=1.23.0",
    "customtkinter>=5.2.0",
    "click>=8.1.0",
    "mistune>=3.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
pdf = ["weasyprint>=60.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pyinstaller>=6.0",
]

[project.scripts]
adm = "src.cli.main:main"

[project.gui-scripts]
adm-gui = "src.gui.main_window:main"

[project.urls]
Homepage = "https://github.com/example/academic-document-manager"
Documentation = "https://github.com/example/academic-document-manager#readme"
Repository = "https://github.com/example/academic-document-manager"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "function1*", "function2*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src", "function1", "function2"]
omit = ["*/tests/*", "*/__pycache__/*"]
